+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # COIN Shell Release Notes
 
+# February 2026
+
+* For CDK projects, check for the use of Yarn workspaces and compile all workspace packages before running CDK commands. This allows CDK code to depend on externalized local packages, such as a "constructs" module.
+  * This feature can be disabled by setting a `COIN_SKIP_WORKSPACE_BUILD` environment variable to "y".
+
 # November 2025
 
 * Added `make help` and `make coin-help` commands that will print out documentation for each command, grouped by category
 * CDK stack synthesizer now defaults to using the current user's IAM credentials to perform deployments instead of trying to assume a role created by the CDK bootstrap.
+* Fix project-specific CDK binary detection so that it can work with yarn/npm workspaces
+* Improve CDK context-clearing behavior to not clear acknowledged-issue-numbers
+* Use latest versions of aws-cdk and aws-cdk-lib
 
 # October 2025